Defined GetRoom, Add GetRoomEmployee/AssignAnimalRoom
Defined method body of GetRoom in Query.
Added GetRoomEmployee to trigger a call to AssignAnimalRoom when database does not currently have a room assignment in the Rooms table.
Refactored DisplayAnimalInfo to now show "Location: Not Found" instead of a null exception when the animal had not been assigned a room.
DisplayAnimalInfoEmployee specifically calls the GetRoomEmployee as used above to ensure all animals are correctly assigned a Room.
Rooms cannot currently be shared between animals due to safety concerns.