Add UpdateShot method chain
Defined UpdateShot along with assisting methods.
First checks if Shot exists. Throws an exception if not found. User can then add the shot as a new shot, re-enter shot name, or cancel the update.
GetShotMenu- Where user decides to Add Shot, Re-enter or cancel.
AddShot- will add a shot if it is not already listed in the database.
GetShot returns the Shot if shotName is found, throws an exception if record not found.
All instances are NOT case sensitive.